fix: 인앱 브라우저 접속시 리다이렉트 설정 (#541)
